+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 - Adds the `isUnderlined` prop to the `Link` component.
 - Adds the `"standalone"` variant to the `Link` component.
 - Adds `NYPL Patterns` to the `Links Accessibility Guide`.
+- Adds the `SocialMediaLinks` component.
 
 ### Updates (Typography Feature Branch)